updates to services,validator, and unit tests for services and controllers

+++ app-server/src/main/java/com/inventory/app/server/service/media/BookService.java
@@ -2,8 +2,10 @@ package com.inventory.app.server.service.media;
 
 import com.inventory.app.server.config.MediaInventoryAdditionalAttributes;
 import com.inventory.app.server.entity.media.Book;
+import com.inventory.app.server.error.NoChangesToUpdateException;
+import com.inventory.app.server.error.ResourceAlreadyExistsException;
+import com.inventory.app.server.error.ResourceNotFoundException;
 import com.inventory.app.server.repository.IBaseDao;
-import com.inventory.app.server.utility.RestPreConditions;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -49,29 +51,33 @@ public class BookService {
     }
 
     public Book create(Book book) {
-        // validations before performing create
-        RestPreConditions.checkAlreadyExists(bookAlreadyExists(book), book);
-
+        if (bookAlreadyExists(book)) {
+            throw new ResourceAlreadyExistsException("Cannot create book because book already exists: " + book);
+        }
         Book bookToSave = cloneBook(book);
         bookToSave.setId(null);
         bookToSave.setVersion(1);
-
         return dao.createOrUpdate(bookToSave);
     }
 
     public Book update(Book updatedBook) {
-        //validations before performing the update
-        Book existingBook = RestPreConditions.checkFound(getBookById(updatedBook.getId()));
-        RestPreConditions.checkEquals(existingBook, updatedBook);
-
+        if (!bookAlreadyExists(updatedBook)) {
+            throw new ResourceNotFoundException("Cannot update book because book does not exist: " + updatedBook);
+        }
+        Book existingBook = getBookById(updatedBook.getId());
+        if (verifyIfBookUpdated(existingBook, updatedBook)) {
+            throw new NoChangesToUpdateException("No updates in book to save. Will not proceed with update. Existing Book: " + existingBook + "Updated Book: " + updatedBook);
+        }
         updatedBook = cloneBook(existingBook, updatedBook);
         updatedBook.setVersion(existingBook.getVersion() + 1);
-
         return dao.createOrUpdate(updatedBook);
     }
 
     public Book deleteById(Long id){
-        Book book = RestPreConditions.checkFound(getBookById(id));
+        Book book = getBookById(id);
+        if (book == null) {
+            throw new ResourceNotFoundException("Cannot delete book because book does not exist.");
+        }
         dao.deleteById(id);
         return book;
     }
@@ -91,5 +97,9 @@ public class BookService {
         return getAllBooksByAuthor(book.getAuthors())
                 .stream()
                 .anyMatch(b -> book.getTitle().equals(b.getTitle()));
+    }
+
+    private boolean verifyIfBookUpdated(Book existingBook, Book updatedBook){
+        return existingBook.equals(updatedBook);
     }
 }
