+++ app-server/src/main/java/com/inventory/app/server/service/media/TelevisionService.java
@@ -2,8 +2,10 @@ package com.inventory.app.server.service.media;
 
 import com.inventory.app.server.config.MediaInventoryAdditionalAttributes;
 import com.inventory.app.server.entity.media.TelevisionShow;
+import com.inventory.app.server.error.NoChangesToUpdateException;
+import com.inventory.app.server.error.ResourceAlreadyExistsException;
+import com.inventory.app.server.error.ResourceNotFoundException;
 import com.inventory.app.server.repository.IBaseDao;
-import com.inventory.app.server.utility.RestPreConditions;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -43,29 +45,34 @@ public class TelevisionService {
     }
 
     public TelevisionShow create(TelevisionShow televisionShow) {
-        // validations before performing create
-        RestPreConditions.checkAlreadyExists(televisionShowAlreadyExists(televisionShow), televisionShow);
-
+        if(televisionShowAlreadyExists(televisionShow)) {
+            throw new ResourceAlreadyExistsException("Cannot create television show because television show already exists: " + televisionShow);
+        }
         TelevisionShow televisionShowToSave = cloneTelevisionShow(televisionShow);
         televisionShowToSave.setId(null);
         televisionShowToSave.setVersion(1);
-
         return dao.createOrUpdate(televisionShowToSave);
     }
 
     public TelevisionShow update(TelevisionShow updatedTelevisionShow) {
-        //validations before performing the update
-        TelevisionShow existingTelevisionShow = RestPreConditions.checkFound(getTelevisionShowById(updatedTelevisionShow.getId()));
-        RestPreConditions.checkEquals(existingTelevisionShow, updatedTelevisionShow);
-
+        if (!televisionShowAlreadyExists(updatedTelevisionShow)) {
+            throw new ResourceNotFoundException("Cannot update television show because television show does not exist: " + updatedTelevisionShow);
+        }
+        TelevisionShow existingTelevisionShow = getTelevisionShowById(updatedTelevisionShow.getId());
+        if (verifyIfTelevisionShowUpdated(existingTelevisionShow, updatedTelevisionShow)) {
+            throw new NoChangesToUpdateException("No updates in television show to save. Will not proceed with update. Existing Television Show: " + existingTelevisionShow + "Updated Television Show: " + updatedTelevisionShow);
+        }
         updatedTelevisionShow = cloneTelevisionShow(existingTelevisionShow, updatedTelevisionShow);
         updatedTelevisionShow.setVersion(existingTelevisionShow.getVersion() + 1);
-
         return dao.createOrUpdate(updatedTelevisionShow);
     }
 
     public TelevisionShow deleteById(Long id){
-        TelevisionShow televisionShow = RestPreConditions.checkFound(getTelevisionShowById(id));
+        TelevisionShow televisionShow = getTelevisionShowById(id);
+        // TODO ADD A NULL CHECK HERE AND THROW EXCEPTION IF TV SHOW DOES NOT EXIST
+        if (televisionShow == null) {
+            throw new ResourceNotFoundException("Cannot delete television show because television show does not exist.");
+        }
         dao.deleteById(id);
         return televisionShow;
     }
@@ -85,5 +92,9 @@ public class TelevisionService {
         return getAllTelevisionShowsByEpisode(televisionShow.getEpisodes())
                 .stream()
                 .anyMatch(t -> televisionShow.getTitle().equals(t.getTitle()));
+    }
+
+    private boolean verifyIfTelevisionShowUpdated(TelevisionShow existingTelevisionShow, TelevisionShow updatedTelevisionShow) {
+        return existingTelevisionShow.equals(updatedTelevisionShow);
     }
 }
